pyproject.toml
[tool.poetry]
name = "theetl"
version = "0.1.0"
description = "Supercustom ETL"
authors = ["Ismael Sanchez <isanchezcasado@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
